refactor(conditions): moves equal conditions into own folder
