+++ src/conditions/equal/equal.ts
